♻️ move provider component into the CommandPalette directory
These components are used in the AppBar component.
