+++ frontend/packages/erd-core/src/features/erd/components/ERDRenderer/CommandPalette/CommandPalette.tsx
@@ -5,9 +5,10 @@ import { DialogDescription, DialogTitle } from '@radix-ui/react-dialog'
 import { Command } from 'cmdk'
 import { type FC, useCallback, useEffect, useState } from 'react'
 import { useTableSelection } from '@/features/erd/hooks'
-import { useCommandPalette, useSchema } from '@/stores'
+import { useSchema } from '@/stores'
 import { TableNode } from '../../ERDContent/components'
 import styles from './CommandPalette.module.css'
+import { useCommandPalette } from './CommandPaletteProvider'
 
 export const CommandPalette: FC = () => {
   const { open, setOpen, toggleOpen } = useCommandPalette()
