+++ frontend/packages/erd-core/src/features/erd/components/ERDRenderer/CommandPalette/index.ts
@@ -1,2 +1,3 @@
 export * from './CommandPalette'
+export { CommandPaletteProvider } from './CommandPaletteProvider'
 export * from './CommandPaletteTriggerButton'
