+++ frontend/packages/erd-core/src/stores/index.ts
@@ -1,3 +1,2 @@
-export * from './commandPalette'
 export * from './schema'
 export * from './userEditing'
